pyproject.toml

[tool.poetry]
name = "co2mon"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
azure-cognitiveservices-vision-computervision = "^0.9.0"
opencv-contrib-python = "^4.5.2"
Pillow = "^8.2.0"
python-dotenv = "^0.17.1"
autopep8 = "^1.5.7"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
